Not introduce new classes for column vector.
