+++ sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OnHeapColumnVector.java
@@ -27,9 +27,116 @@ import org.apache.spark.unsafe.Platform;
 /**
  * A column backed by an in memory JVM array.
  */
-public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
-  public OnHeapColumnVector(int capacity, DataType type) {
-    super(capacity, capacity, type, false);
+public final class OnHeapColumnVector extends ColumnVector {
+  protected OnHeapColumnVector(int capacity, int childCapacity, DataType type, boolean isConstant) {
+    super(capacity, childCapacity, type, MemoryMode.ON_HEAP, isConstant);
+
+    this.isConstant = isConstant;
+    reserveInternal(capacity);
+    reset();
+  }
+
+  private static final boolean bigEndianPlatform =
+    ByteOrder.nativeOrder().equals(ByteOrder.BIG_ENDIAN);
+
+  // The data stored in these arrays need to maintain binary compatible. We can
+  // directly pass this buffer to external components.
+
+  // This is faster than a boolean array and we optimize this over memory footprint.
+  private byte[] nulls;
+
+  // Array for each type. Only 1 is populated for any type.
+  private byte[] byteData;
+  private short[] shortData;
+  private int[] intData;
+  private long[] longData;
+  private float[] floatData;
+  private double[] doubleData;
+
+  // Only set if type is Array.
+  private int[] arrayLengths;
+  private int[] arrayOffsets;
+
+
+  @Override
+  public long valuesNativeAddress() {
+    throw new RuntimeException("Cannot get native address for on heap column");
+  }
+  @Override
+  public long nullsNativeAddress() {
+    throw new RuntimeException("Cannot get native address for on heap column");
+  }
+
+  @Override
+  public void close() {
+  }
+
+  // Spilt this function out since it is the slow path.
+  protected void reserveInternal(int newCapacity) {
+    if (this.resultArray != null || DecimalType.isByteArrayDecimalType(type)) {
+      int[] newLengths = new int[newCapacity];
+      int[] newOffsets = new int[newCapacity];
+      if (this.arrayLengths != null) {
+        System.arraycopy(this.arrayLengths, 0, newLengths, 0, elementsAppended);
+        System.arraycopy(this.arrayOffsets, 0, newOffsets, 0, elementsAppended);
+      }
+      arrayLengths = newLengths;
+      arrayOffsets = newOffsets;
+    } else if (type instanceof BooleanType) {
+      if (byteData == null || byteData.length < newCapacity) {
+        byte[] newData = new byte[newCapacity];
+        if (byteData != null) System.arraycopy(byteData, 0, newData, 0, elementsAppended);
+        byteData = newData;
+      }
+    } else if (type instanceof ByteType) {
+      if (byteData == null || byteData.length < newCapacity) {
+        byte[] newData = new byte[newCapacity];
+        if (byteData != null) System.arraycopy(byteData, 0, newData, 0, elementsAppended);
+        byteData = newData;
+      }
+    } else if (type instanceof ShortType) {
+      if (shortData == null || shortData.length < newCapacity) {
+        short[] newData = new short[newCapacity];
+        if (shortData != null) System.arraycopy(shortData, 0, newData, 0, elementsAppended);
+        shortData = newData;
+      }
+    } else if (type instanceof IntegerType || type instanceof DateType ||
+      DecimalType.is32BitDecimalType(type)) {
+      if (intData == null || intData.length < newCapacity) {
+        int[] newData = new int[newCapacity];
+        if (intData != null) System.arraycopy(intData, 0, newData, 0, elementsAppended);
+        intData = newData;
+      }
+    } else if (type instanceof LongType || type instanceof TimestampType ||
+        DecimalType.is64BitDecimalType(type)) {
+      if (longData == null || longData.length < newCapacity) {
+        long[] newData = new long[newCapacity];
+        if (longData != null) System.arraycopy(longData, 0, newData, 0, elementsAppended);
+        longData = newData;
+      }
+    } else if (type instanceof FloatType) {
+      if (floatData == null || floatData.length < newCapacity) {
+        float[] newData = new float[newCapacity];
+        if (floatData != null) System.arraycopy(floatData, 0, newData, 0, elementsAppended);
+        floatData = newData;
+      }
+    } else if (type instanceof DoubleType) {
+      if (doubleData == null || doubleData.length < newCapacity) {
+        double[] newData = new double[newCapacity];
+        if (doubleData != null) System.arraycopy(doubleData, 0, newData, 0, elementsAppended);
+        doubleData = newData;
+      }
+    } else if (resultStruct != null) {
+      // Nothing to store.
+    } else {
+      throw new RuntimeException("Unhandled " + type);
+    }
+
+    byte[] newNulls = new byte[newCapacity];
+    if (nulls != null) System.arraycopy(nulls, 0, newNulls, 0, elementsAppended);
+    nulls = newNulls;
+
+    capacity = newCapacity;
   }
 
   //
@@ -38,20 +145,32 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
 
   @Override
   public void putNotNull(int rowId) {
-    nulls[rowId] = (byte)0;
+    if (isConstant) {
+      nulls[0] = (byte)0;
+    } else {
+      nulls[rowId] = (byte)0;
+    }
   }
 
   @Override
   public void putNull(int rowId) {
-    nulls[rowId] = (byte)1;
+    if (isConstant) {
+      nulls[0] = (byte)1;
+    } else {
+      nulls[rowId] = (byte)1;
+    }
     ++numNulls;
     anyNullsSet = true;
   }
 
   @Override
   public void putNulls(int rowId, int count) {
-    for (int i = 0; i < count; ++i) {
-      nulls[rowId + i] = (byte)1;
+    if (isConstant) {
+      nulls[0] = (byte)1;
+    } else {
+      for (int i = 0; i < count; ++i) {
+        nulls[rowId + i] = (byte)1;
+      }
     }
     anyNullsSet = true;
     numNulls += count;
@@ -60,14 +179,22 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
   @Override
   public void putNotNulls(int rowId, int count) {
     if (!anyNullsSet) return;
-    for (int i = 0; i < count; ++i) {
-      nulls[rowId + i] = (byte)0;
+    if (isConstant) {
+      nulls[0] = (byte)0;
+    } else {
+      for (int i = 0; i < count; ++i) {
+        nulls[rowId + i] = (byte)0;
+      }
     }
   }
 
   @Override
   public boolean isNullAt(int rowId) {
-    return nulls[rowId] == 1;
+    if (isConstant) {
+      return nulls[0] == 1;
+    } else {
+      return nulls[rowId] == 1;
+    }
   }
 
   //
@@ -76,20 +203,32 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
 
   @Override
   public void putBoolean(int rowId, boolean value) {
-    byteData[rowId] = (byte)((value) ? 1 : 0);
+    if (isConstant) {
+      byteData[0] = (byte)((value) ? 1 : 0);
+    } else {
+      byteData[rowId] = (byte)((value) ? 1 : 0);
+    }
   }
 
   @Override
   public void putBooleans(int rowId, int count, boolean value) {
     byte v = (byte)((value) ? 1 : 0);
-    for (int i = 0; i < count; ++i) {
-      byteData[i + rowId] = v;
+    if (isConstant) {
+      byteData[0] = v;
+    } else {
+      for (int i = 0; i < count; ++i) {
+        byteData[i + rowId] = v;
+      }
     }
   }
 
   @Override
   public boolean getBoolean(int rowId) {
-    return byteData[rowId] == 1;
+    if (isConstant) {
+      return byteData[0] == 1;
+    } else {
+      return byteData[rowId] == 1;
+    }
   }
 
   //
@@ -100,27 +239,47 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
 
   @Override
   public void putByte(int rowId, byte value) {
-    byteData[rowId] = value;
+    if (isConstant) {
+      byteData[0] = value;
+    } else {
+      byteData[rowId] = value;
+    }
   }
 
   @Override
   public void putBytes(int rowId, int count, byte value) {
-    for (int i = 0; i < count; ++i) {
-      byteData[i + rowId] = value;
+    if (isConstant) {
+      byteData[0] = value;
+    } else {
+      for (int i = 0; i < count; ++i) {
+        byteData[i + rowId] = value;
+      }
     }
   }
 
   @Override
   public void putBytes(int rowId, int count, byte[] src, int srcIndex) {
-    System.arraycopy(src, srcIndex, byteData, rowId, count);
+    if (isConstant) {
+      System.arraycopy(src, srcIndex, byteData, 0, 1);
+    } else {
+      System.arraycopy(src, srcIndex, byteData, rowId, count);
+    }
   }
 
   @Override
   public byte getByte(int rowId) {
     if (dictionary == null) {
-      return byteData[rowId];
+      if (isConstant) {
+        return byteData[0];
+      } else {
+        return byteData[rowId];
+      }
     } else {
-      return (byte) dictionary.decodeToInt(dictionaryIds.getInt(rowId));
+      if (isConstant) {
+        return (byte) dictionary.decodeToInt(dictionaryIds.getInt(0));
+      } else {
+        return (byte) dictionary.decodeToInt(dictionaryIds.getInt(rowId));
+      }
     }
   }
 
@@ -130,27 +289,47 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
 
   @Override
   public void putShort(int rowId, short value) {
-    shortData[rowId] = value;
+    if (isConstant) {
+      shortData[0] = value;
+    } else {
+      shortData[rowId] = value;
+    }
   }
 
   @Override
   public void putShorts(int rowId, int count, short value) {
-    for (int i = 0; i < count; ++i) {
-      shortData[i + rowId] = value;
+    if (isConstant) {
+      shortData[0] = value;
+    } else {
+      for (int i = 0; i < count; ++i) {
+        shortData[i + rowId] = value;
+      }
     }
   }
 
   @Override
   public void putShorts(int rowId, int count, short[] src, int srcIndex) {
-    System.arraycopy(src, srcIndex, shortData, rowId, count);
+    if (isConstant) {
+      System.arraycopy(src, srcIndex, shortData, 0, 1);
+    } else {
+      System.arraycopy(src, srcIndex, shortData, rowId, count);
+    }
   }
 
   @Override
   public short getShort(int rowId) {
     if (dictionary == null) {
-      return shortData[rowId];
+      if (isConstant) {
+        return shortData[0];
+      } else {
+        return shortData[rowId];
+      }
     } else {
-      return (short) dictionary.decodeToInt(dictionaryIds.getInt(rowId));
+      if (isConstant) {
+        return (short) dictionary.decodeToInt(dictionaryIds.getInt(0));
+      } else {
+        return (short) dictionary.decodeToInt(dictionaryIds.getInt(rowId));
+      }
     }
   }
 
@@ -161,28 +340,47 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
 
   @Override
   public void putInt(int rowId, int value) {
-    intData[rowId] = value;
+    if (isConstant) {
+      intData[0] = value;
+    } else {
+      intData[rowId] = value;
+    }
   }
 
   @Override
   public void putInts(int rowId, int count, int value) {
-    for (int i = 0; i < count; ++i) {
-      intData[i + rowId] = value;
+    if (isConstant) {
+      intData[0] = value;
+    } else {
+      for (int i = 0; i < count; ++i) {
+        intData[i + rowId] = value;
+      }
     }
   }
 
   @Override
   public void putInts(int rowId, int count, int[] src, int srcIndex) {
-    System.arraycopy(src, srcIndex, intData, rowId, count);
+    if (isConstant) {
+      System.arraycopy(src, srcIndex, intData, 0, 1);
+    } else {
+      System.arraycopy(src, srcIndex, intData, rowId, count);
+    }
   }
 
   @Override
   public void putIntsLittleEndian(int rowId, int count, byte[] src, int srcIndex) {
     int srcOffset = srcIndex + Platform.BYTE_ARRAY_OFFSET;
-    for (int i = 0; i < count; ++i, srcOffset += 4) {
-      intData[i + rowId] = Platform.getInt(src, srcOffset);
+    if (isConstant) {
+      intData[0] = Platform.getInt(src, srcOffset);
       if (bigEndianPlatform) {
-        intData[i + rowId] = java.lang.Integer.reverseBytes(intData[i + rowId]);
+        intData[0] = java.lang.Integer.reverseBytes(intData[0]);
+      }
+    } else {
+      for (int i = 0; i < count; ++i, srcOffset += 4) {
+        intData[i + rowId] = Platform.getInt(src, srcOffset);
+        if (bigEndianPlatform) {
+          intData[i + rowId] = java.lang.Integer.reverseBytes(intData[i + rowId]);
+        }
       }
     }
   }
@@ -190,9 +388,17 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
   @Override
   public int getInt(int rowId) {
     if (dictionary == null) {
-      return intData[rowId];
+      if (isConstant) {
+        return intData[0];
+      } else {
+        return intData[rowId];
+      }
     } else {
-      return dictionary.decodeToInt(dictionaryIds.getInt(rowId));
+      if (isConstant) {
+        return dictionary.decodeToInt(dictionaryIds.getInt(0));
+      } else {
+        return dictionary.decodeToInt(dictionaryIds.getInt(rowId));
+      }
     }
   }
 
@@ -202,28 +408,47 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
 
   @Override
   public void putLong(int rowId, long value) {
-    longData[rowId] = value;
+    if (isConstant) {
+      longData[0] = value;
+    } else {
+      longData[rowId] = value;
+    }
   }
 
   @Override
   public void putLongs(int rowId, int count, long value) {
-    for (int i = 0; i < count; ++i) {
-      longData[i + rowId] = value;
+    if (isConstant) {
+      longData[0] = value;
+    } else {
+      for (int i = 0; i < count; ++i) {
+        longData[i + rowId] = value;
+      }
     }
   }
 
   @Override
   public void putLongs(int rowId, int count, long[] src, int srcIndex) {
-    System.arraycopy(src, srcIndex, longData, rowId, count);
+    if (isConstant) {
+      System.arraycopy(src, srcIndex, longData, 0, 1);
+    } else {
+      System.arraycopy(src, srcIndex, longData, rowId, count);
+    }
   }
 
   @Override
   public void putLongsLittleEndian(int rowId, int count, byte[] src, int srcIndex) {
     int srcOffset = srcIndex + Platform.BYTE_ARRAY_OFFSET;
-    for (int i = 0; i < count; ++i, srcOffset += 8) {
-      longData[i + rowId] = Platform.getLong(src, srcOffset);
+    if (isConstant) {
+      longData[0] = Platform.getLong(src, srcOffset);
       if (bigEndianPlatform) {
-        longData[i + rowId] = java.lang.Long.reverseBytes(longData[i + rowId]);
+        longData[0] = java.lang.Long.reverseBytes(longData[0]);
+      }
+    } else {
+      for (int i = 0; i < count; ++i, srcOffset += 8) {
+        longData[i + rowId] = Platform.getLong(src, srcOffset);
+        if (bigEndianPlatform) {
+          longData[i + rowId] = java.lang.Long.reverseBytes(longData[i + rowId]);
+        }
       }
     }
   }
@@ -231,9 +456,17 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
   @Override
   public long getLong(int rowId) {
     if (dictionary == null) {
-      return longData[rowId];
+      if (isConstant) {
+        return longData[0];
+      } else {
+        return longData[rowId];
+      }
     } else {
-      return dictionary.decodeToLong(dictionaryIds.getInt(rowId));
+      if (isConstant) {
+        return dictionary.decodeToLong(dictionaryIds.getInt(0));
+      } else {
+        return dictionary.decodeToLong(dictionaryIds.getInt(rowId));
+      }
     }
   }
 
@@ -243,28 +476,49 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
 
   @Override
   public void putFloat(int rowId, float value) {
-    floatData[rowId] = value;
+    if (isConstant) {
+      floatData[0] = value;
+    } else {
+      floatData[rowId] = value;
+    }
   }
 
   @Override
   public void putFloats(int rowId, int count, float value) {
-    Arrays.fill(floatData, rowId, rowId + count, value);
+    if (isConstant) {
+      Arrays.fill(floatData, 0, 1, value);
+    } else {
+      Arrays.fill(floatData, rowId, rowId + count, value);
+    }
   }
 
   @Override
   public void putFloats(int rowId, int count, float[] src, int srcIndex) {
-    System.arraycopy(src, srcIndex, floatData, rowId, count);
+    if (isConstant) {
+      System.arraycopy(src, srcIndex, floatData, 0, 1);
+    } else {
+      System.arraycopy(src, srcIndex, floatData, rowId, count);
+    }
   }
 
   @Override
   public void putFloats(int rowId, int count, byte[] src, int srcIndex) {
     if (!bigEndianPlatform) {
-      Platform.copyMemory(src, Platform.BYTE_ARRAY_OFFSET + srcIndex, floatData,
+      if (isConstant) {
+        Platform.copyMemory(src, Platform.BYTE_ARRAY_OFFSET + srcIndex, floatData,
+          Platform.DOUBLE_ARRAY_OFFSET, 4);
+      } else {
+        Platform.copyMemory(src, Platform.BYTE_ARRAY_OFFSET + srcIndex, floatData,
           Platform.DOUBLE_ARRAY_OFFSET + rowId * 4, count * 4);
+      }
     } else {
       ByteBuffer bb = ByteBuffer.wrap(src).order(ByteOrder.LITTLE_ENDIAN);
-      for (int i = 0; i < count; ++i) {
-        floatData[i + rowId] = bb.getFloat(srcIndex + (4 * i));
+      if (isConstant) {
+        floatData[0] = bb.getFloat(srcIndex);
+      } else {
+        for (int i = 0; i < count; ++i) {
+          floatData[i + rowId] = bb.getFloat(srcIndex + (4 * i));
+        }
       }
     }
   }
@@ -272,9 +526,17 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
   @Override
   public float getFloat(int rowId) {
     if (dictionary == null) {
-      return floatData[rowId];
+      if (isConstant) {
+        return floatData[0];
+      } else {
+        return floatData[rowId];
+      }
     } else {
-      return dictionary.decodeToFloat(dictionaryIds.getInt(rowId));
+      if (isConstant) {
+        return dictionary.decodeToFloat(dictionaryIds.getInt(0));
+      } else {
+        return dictionary.decodeToFloat(dictionaryIds.getInt(rowId));
+      }
     }
   }
 
@@ -284,28 +546,49 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
 
   @Override
   public void putDouble(int rowId, double value) {
-    doubleData[rowId] = value;
+    if (isConstant) {
+      doubleData[0] = value;
+    } else {
+      doubleData[rowId] = value;
+    }
   }
 
   @Override
   public void putDoubles(int rowId, int count, double value) {
-    Arrays.fill(doubleData, rowId, rowId + count, value);
+    if (isConstant) {
+      Arrays.fill(doubleData, 0, 1, value);
+    } else {
+      Arrays.fill(doubleData, rowId, rowId + count, value);
+    }
   }
 
   @Override
   public void putDoubles(int rowId, int count, double[] src, int srcIndex) {
-    System.arraycopy(src, srcIndex, doubleData, rowId, count);
+    if (isConstant) {
+      System.arraycopy(src, srcIndex, doubleData, 0, 1);
+    } else {
+      System.arraycopy(src, srcIndex, doubleData, rowId, count);
+    }
   }
 
   @Override
   public void putDoubles(int rowId, int count, byte[] src, int srcIndex) {
     if (!bigEndianPlatform) {
-      Platform.copyMemory(src, Platform.BYTE_ARRAY_OFFSET + srcIndex, doubleData,
+      if (isConstant) {
+        Platform.copyMemory(src, Platform.BYTE_ARRAY_OFFSET + srcIndex, doubleData,
+          Platform.DOUBLE_ARRAY_OFFSET, 8);
+      } else {
+        Platform.copyMemory(src, Platform.BYTE_ARRAY_OFFSET + srcIndex, doubleData,
           Platform.DOUBLE_ARRAY_OFFSET + rowId * 8, count * 8);
+      }
     } else {
       ByteBuffer bb = ByteBuffer.wrap(src).order(ByteOrder.LITTLE_ENDIAN);
-      for (int i = 0; i < count; ++i) {
-        doubleData[i + rowId] = bb.getDouble(srcIndex + (8 * i));
+      if (isConstant) {
+        doubleData[0] = bb.getDouble(srcIndex);
+      } else { 
+        for (int i = 0; i < count; ++i) {
+          doubleData[i + rowId] = bb.getDouble(srcIndex + (8 * i));
+        }
       }
     }
   }
@@ -313,9 +596,17 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
   @Override
   public double getDouble(int rowId) {
     if (dictionary == null) {
-      return doubleData[rowId];
+      if (isConstant) {
+        return doubleData[0];
+      } else {
+        return doubleData[rowId];
+      }
     } else {
-      return dictionary.decodeToDouble(dictionaryIds.getInt(rowId));
+      if (isConstant) {
+        return dictionary.decodeToDouble(dictionaryIds.getInt(0));
+      } else {
+        return dictionary.decodeToDouble(dictionaryIds.getInt(rowId));
+      }
     }
   }
 
@@ -325,17 +616,30 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
 
   @Override
   public int getArrayLength(int rowId) {
-    return arrayLengths[rowId];
+    if (isConstant) {
+      return arrayLengths[0];
+    } else {
+      return arrayLengths[rowId];
+    }
   }
   @Override
   public int getArrayOffset(int rowId) {
-    return arrayOffsets[rowId];
+    if (isConstant) {
+      return arrayOffsets[0];
+    } else {
+      return arrayOffsets[rowId];
+    }
   }
 
   @Override
   public void putArray(int rowId, int offset, int length) {
-    arrayOffsets[rowId] = offset;
-    arrayLengths[rowId] = length;
+    if (isConstant) {
+      arrayOffsets[0] = offset;
+      arrayLengths[0] = length;
+    } else {
+      arrayOffsets[rowId] = offset;
+      arrayLengths[rowId] = length;
+    }
   }
 
   @Override
@@ -351,8 +655,13 @@ public final class OnHeapColumnVector extends OnHeapColumnVectorBase {
   @Override
   public int putByteArray(int rowId, byte[] value, int offset, int length) {
     int result = arrayData().appendBytes(length, value, offset);
-    arrayOffsets[rowId] = result;
-    arrayLengths[rowId] = length;
+    if (isConstant) {
+      arrayOffsets[0] = result;
+      arrayLengths[0] = length;
+    } else {
+      arrayOffsets[rowId] = result;
+      arrayLengths[rowId] = length;
+    }
     return result;
   }
